==> 03 SVL.ipynb <==
import glob
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import calendar
from sklearn import preprocessing
from sklearn.cluster import KMeans
import matplotlib.style as style
from contextlib import contextmanager
from sklearn.model_selection import train_test_split
from datetime import datetime
from sklearn import metrics
from sklearn.metrics import pairwise_distances
from sklearn.metrics import davies_bouldin_score
from sklearn import metrics
from sklearn.metrics import pairwise_distances
import math
import os
import numpy as np
import warnings
warnings.filterwarnings("ignore")
###############################################################################################################################
#####################################################  MAESTRAS  ##############################################################
###############################################################################################################################

os.chdir('D:/Reporting/Maestras/')
Maestro_SKU = pd.read_excel('Materiales  Abi.xlsx', sheet_name='Hoja1')
Maestro_SKU.rename(columns = {'Material': 'Sku_KOA','Material ABI': 'Sku','HL/cj':'HL_cj'} ,inplace=True)

os.chdir('D:/Reporting/Maestras/')
Maestro_Clientes = pd.read_excel('Maestra de clientes.xlsx', sheet_name='Hoja1')
Maestro_Clientes.rename(columns = {'Cod Cliente': 'COD_ABI','Codigo Politica': 'COD_Politica'} ,inplace=True)

###############################################################################################################################
#####################################################  RECORTES  ##############################################################
###############################################################################################################################

#os.chdir('D:/Reporting/Bases Manuales/13 SERVICE LEVEL/Recortes')
#Recortes = []
#csv_files = glob.glob('*.txt')

#for filename in csv_files:
  #  data = pd.read_csv(filename,encoding='ANSI', sep=',')
  #  #data['filename']=filename
   # Recortes.append(data)
#Recortes = pd.concat(Recortes,ignore_index=True)

#Recortes = Recortes[["F.Pedido ","Client","DPS","Motivo    "," C.Prd", "Cant"]]
#Recortes.rename(columns = {'F.Pedido ': 'Fecha_t','DPS':'COD_CD','Motivo    ':'Motivos',' C.Prd':'Sku','Cant':'Cj','Client':'COD_ABI'} ,inplace=True)

#os.chdir('D:/Reporting/Maestras/')
#Matriz_Reco = pd.read_excel('MATRIZ_RECORTES.xlsx', sheet_name='RECORTES')
#Recortes = pd.merge(Recortes,Matriz_Reco,on=["Motivos"],how="outer")

#Maestra_Clientes_S = Maestro_Clientes[["COD_ABI","Troncal"]]
#Recortes = pd.merge(Recortes,Maestra_Clientes_S,on=["COD_ABI"],how="outer")
#Recortes = Recortes.dropna(subset=['Fecha_t'])
#Recortes['Data'] = "Recortes"
#Recortes.Cj = Recortes.Cj.astype(int)
#Recortes.Cj =Recortes.Cj/100
#Recortes.Cj = Recortes.Cj.round(2)
#Recortes.Sku =Recortes.Sku*10

#Recortes.to_csv("D:/Reporting/ETL/04 Service Level/Reco_SL.txt")

###############################################################################################################################
#####################################################   SELL   ################################################################
###############################################################################################################################

os.chdir('D:/Reporting/Bases Manuales/02 SELL/Sell IN')
Sell_in = []
csv_files = glob.glob('*.txt')

for filename in csv_files:
    data = pd.read_csv(filename,encoding='UTF-8', sep='\t')
    data['filename']=filename
    Sell_in.append(data)
Sell_in = pd.concat(Sell_in,ignore_index=True)


Sell_SL = Sell_in[["Año Mes Dia Cierre","C.Cli","C.Dep.Ori","COD Motivo Movimiento","C.Prd", "#Cj Vta Bruta SUM"]]
Sell_SL.rename(columns = {'Año Mes Dia Cierre': 'Fecha_t','C.Cli':'COD_ABI' ,'C.Dep.Ori':'COD_CD',' C.Prd':'Sku','#Cj Vta Bruta SUM':'Cj'} ,inplace=True)

os.chdir('D:/Reporting/Maestras/')
Matriz_Rech = pd.read_excel('Matriz_Rechazo.xlsx', sheet_name='Hoja1')

Sell_SL = pd.merge(Sell_SL,Matriz_Rech,on=["COD Motivo Movimiento"],how="outer")
Sell_SL = Sell_SL.dropna(subset=['Fecha_t'])

Sell_SL.rename(columns = {'DSC Motivo Movimiento':'Motivos','C.Prd':'Sku'} ,inplace=True)
#Sell_SL = Sell_SL[["Fecha_t","COD_ABI","COD_CD","Motivos","Sku", "Cj","Tipo Recorte"]]

Maestra_Clientes_S = Maestro_Clientes[["COD_ABI","Troncal"]]
Sell_SL = pd.merge(Sell_SL,Maestra_Clientes_S,on=["COD_ABI"],how="outer")
#Sell_SL.Cj = Sell_SL.Cj.round(2)
Sell_SL = Sell_SL[Sell_SL["Cj"]!=0]
Sell_SL['Data']="Sell"
Sell_SL.rename(columns = {'Tipo Recorte':'Tipo_Recorte'} ,inplace=True)
Sell_SL = Sell_SL.dropna(subset=['Fecha_t'])

Rechazo_Sell = Sell_SL[Sell_SL.Tipo_Recorte.isin(["Logistico","Comercial"])]
#planets[planets.year.isin([2008, 2009]) & planets.mass.isnull()]
Rechazo_Sell['Data']="Rechazo"
Rechazo_Sell.rename(columns = {'Tipo_Recorte':'Tipo Recorte'} ,inplace=True)

Sell_SL =Sell_SL.append(Rechazo_Sell)
SL =Sell_SL#.append(Recortes)

Sell_SL.to_csv("D:/Reporting/ETL/04 Service Level/Sell_SL.txt")

###############################################################################################################################
####################################################   POLITICA   #############################################################
###############################################################################################################################

os.chdir('D:/Reporting/Bases Manuales/03 POLITICA/Diaria')
Politica = []
xx=[]
csv_files = glob.glob('*.xlsx')

for filename in csv_files:
    data = pd.read_excel(filename,sheet_name='Hoja1')#, dtype ={'FCST HL ajustado':np.float64,'Stock Total Hl':np.float64, 'Proyectado':np.float64})
    data['filename']=filename
    xx=data.filename.str.split(".", n=1, expand=True)
    data['Fecha_t']=xx[0]
    Politica.append(data)
Politica = pd.concat(Politica,ignore_index=True)

Pol_SL = Politica[["Region","Cod. Producto","FCST HL ajustado","Stock Total Hl","Fecha_t","Proyectado"]]
Pol_SL.rename(columns = {'Cod. Producto': 'Sku','FCST HL ajustado':'FCST','Stock Total Hl':'Stock','Region':'COD_Politica'} ,inplace=True)
#Pol_SL = Pol_SL[Pol_SL["COD_Politica"] != "Puente Alto"]

##########################################################################################################################


os.chdir('D:/Reporting/Bases Manuales/03 POLITICA/Nuevo formato')
Politica2 = []
csv_files = glob.glob('*.xlsx')

for filename in csv_files:
    data = pd.read_excel(filename,sheet_name='Hoja1')#, dtype={'FCST_HL':np.float64,'Seg_HL':np.float64, 'Repo_HL':np.float64, 'Max_HL':np.float64, 'OOR_OUT_HL':np.float64, 'OOR_OVER_HL':np.float64} ,decimal=',')
    data['filename']=filename
    Politica2.append(data)
Politica2 = pd.concat(Politica2,ignore_index=True)

Titulo = Politica2[["filename"]]
Titulo = Titulo["filename"].str.split(".", n=1, expand=True)
Politica2["Fecha_t"]=Titulo[0].str.replace(r".","")

Politica2.Seg_HL = np.round(Politica2.Seg_HL,1)#Seg_HL
Politica2.Repo_HL = np.round(Politica2.Repo_HL,1)#Repo_HL
Politica2.Max_HL = np.round(Politica2.Max_HL,1)#Max_HL
Politica2.OOR_OUT_HL = np.round(Politica2.OOR_OUT_HL,1)#OOR_OUT_HL
Politica2.OOR_OVER_HL = np.round(Politica2.OOR_OVER_HL,1)#OOR_OVER_HL

Politica2.D0_PLL = np.round(Politica2.D0_PLL,1)#D0_PLL
Politica2.Transito_PLL = np.round(Politica2.Transito_PLL,1)#Transito_PLL
Politica2.D1_PLL = np.round(Politica2.D1_PLL,1)#D1_PLL
Politica2.SugT_HL = np.round(Politica2.SugT_HL,1)#SugT_HL
Politica2.SugR_HL = np.round(Politica2.SugR_HL,1)#SugR_HL
Politica2.Dispo_HL = np.round(Politica2.Dispo_HL,1)#Dispo_HL
Politica2.Stock_Py_HL = np.round(Politica2.Stock_Py_HL,1)#Stock_Py_HL
Politica2.Stock_Tron_HL = np.round(Politica2.Stock_Tron_HL,1)#Stock_Tron_HL
Politica2.FCST_HL = np.round(Politica2.FCST_HL,1)#FCST_HL
Politica2.Dispo_ABI_HL = np.round(Politica2.Dispo_ABI_HL,1)#Dispo_ABI_HL
Politica2.Dispo_Centro_HL = np.round(Politica2.Dispo_Centro_HL,1)# Dispo_Centro_HL 

Politica2 = Politica2[['Fecha_t', 'Region','Cod. Producto','FCST_ajustado_HL','Stock_Tron_HL', 'Stock_Py_HL']]
Politica2.rename(columns ={'FCST_ajustado_HL':'FCST', 'Stock_Py_HL':'Proyectado', 'Stock_Tron_HL':'Stock', 'Region':'COD_Politica','Cod. Producto': 'Sku'},inplace=True)
Politica2 = Politica2.fillna(0)

Pol_SL = Pol_SL.append(Politica2)

###############################################################################################################

Pol_SL.FCST = Pol_SL.FCST.round(2)
Pol_SL.Stock = Pol_SL.Stock.round(2)
Pol_SL.Proyectado = Pol_SL.Proyectado.round(2)
Pol_SL['Stock'] = np.where(Pol_SL['Stock']<0, 0,Pol_SL['Stock'])

#### calculo#####
Pol_SL['St_out_t'] = (Pol_SL.FCST/25) - (Pol_SL.Stock.fillna(0) + Pol_SL.Proyectado.fillna(0))
Pol_SL = Pol_SL[Pol_SL["St_out_t"]>0]

Pol_SL.COD_Politica = Pol_SL.COD_Politica.replace("Centro", "Carval")

Maestra_Clientes_S = Maestro_Clientes[["COD_Politica","Troncal","COD_ABI"]]
Pol_SL = pd.merge(Pol_SL,Maestra_Clientes_S,on=["COD_Politica"],how="outer")
Pol_SL = Pol_SL.dropna(subset=['Fecha_t'])

Pol_SL['Motivos']="SO_TEO"
Pol_SL['Tipo Recorte']="SO_TEO"
Pol_SL['COD_CD']="40"
Pol_SL['Data']="SO_TEO"

Pol_SL = Pol_SL[Pol_SL["St_out_t"]>0] 
Pol_SL = Pol_SL[["Fecha_t","COD_CD","COD_ABI","Sku","St_out_t","Motivos","Tipo Recorte","Data","Troncal","COD_Politica"]]
Pol_SL.St_out_t = Pol_SL.St_out_t.round(2)

Pol_SL.to_csv("D:/Reporting/ETL/04 Service Level/Pol_SL.txt")

##############################   3 DV   ##############################

DV3 = Politica[["Region","Cod. Producto","FCST HL ajustado","Stock Total Hl","Fecha_t","Proyectado"]]
DV3.rename(columns = {'Cod. Producto': 'Sku','FCST HL ajustado':'FCST','Stock Total Hl':'Stock','Region':'COD_Politica'} ,inplace=True)
#DV3 = DV3[DV3["COD_Politica"] != "Puente Alto"]
DV3.COD_Politica = DV3.COD_Politica.replace("Centro", "Carval")

DV3.FCST = DV3.FCST.round(1).fillna(0)
DV3.Stock = DV3.Stock.round(1).fillna(0)
DV3.Proyectado = DV3.Proyectado.round(1).fillna(0)

Maestra_Clientes_S = Maestro_Clientes[["COD_Politica","Troncal","COD_ABI"]]
DV3 = pd.merge(DV3,Maestra_Clientes_S,on=["COD_Politica"],how="outer")
DV3 = DV3.dropna(subset=['Fecha_t'])

DV3['Motivos']="SO_TEO"
DV3['Tipo Recorte']="SO_TEO"
DV3['Data']="DV3"
DV3['COD_CD']="40"

DV3['FCST_lin'] = DV3.FCST*(3/25)
DV3['FCST_lin1'] = DV3.FCST*(1/25)
DV3.FCST_lin1 = DV3.FCST_lin1.astype(float).fillna(0).round(1)
DV3.FCST_lin = DV3.FCST_lin.astype(float).fillna(0).round(1)
DV3['Stock'] = np.where(DV3['Stock']<0, 0,DV3['Stock'])

DV3['St_out_t'] = np.where((DV3.FCST_lin - DV3.Stock.fillna(0))<1, 0,(DV3.FCST_lin1)).round(1).astype(float)
DV3.FCST_lin = DV3.FCST_lin.round(2)

DV3 = DV3[DV3["St_out_t"]>0]                               
DV3.St_out_t = DV3.St_out_t.round(2)

DV3.to_csv("D:/Reporting/ETL/04 Service Level/Pol_SL3Dv.txt")
DV3 = DV3[["Fecha_t","COD_CD","COD_ABI","Sku","St_out_t","Motivos","Tipo Recorte","Data","Troncal","COD_Politica"]]



Pol_SL =Pol_SL.append(DV3)

###############################################################################################################################
#################################################   RECHAZO BOTTLERS   ########################################################
###############################################################################################################################

#os.chdir('D:/Reporting/Bases Manuales/13 SERVICE LEVEL/Rechazo Bottlers')
#Rech_Bott = []
#csv_files = glob.glob('*.xlsx')

#for filename in csv_files:
    #data = pd.read_excel(filename,sheet_name='Hoja1')
    #data['filename']=filename
    #Rech_Bott.append(data)
#Rech_Bott = pd.concat(Rech_Bott,ignore_index=True)
#Rech_Bott = Rech_Bott[["Fecha","N° planilla","Origen","SKU","Cantidad"]]

#Rech_Bott.rename(columns = {'SKU': 'Sku','Origen':'COD_ABI','Fecha':'Fecha_t','Cantidad':'Cj'} ,inplace=True)
#Rech_Bott.Cj = Rech_Bott.Cj.round(2)

#Maestra_Clientes_S = Maestro_Clientes[["COD_ABI","Troncal"]]
#Rech_Bott = pd.merge(Rech_Bott,Maestra_Clientes_S,on=["COD_ABI"],how="outer")
#
#Rech_Bott['Motivos']="Rechazo Bottler"
#Rech_Bott['Tipo Recorte']="Rechazo Bottler"
#Rech_Bott['Data'] = "Rechazo Bottler"
#Rech_Bott = Rech_Bott.dropna(subset=['Sku'])


###############################################################################################################################
##########################################################   SL   #############################################################
###############################################################################################################################

SL =SL.append(Pol_SL)
#SL =SL.append(Rech_Bott)
SL.St_out_t = SL.St_out_t.round(2)
SL['    Client']="0"
SL['Client']="0"
#Maestra_Clientes_S = Maestro_Clientes[["COD_ABI","Troncal"]]
#SL = pd.merge(SL,Maestra_Clientes_S,on=["COD_ABI"],how="outer")
SL
SL.to_csv("D:/Reporting/ETL/04 Service Level/sl1.txt")



############################################# SL DEL BOTTLER ##################################################################
#os.chdir('D:/Reporting/Bases Manuales/13 SERVICE LEVEL/Service Level Bottler/DATA')
#SL_KOE = []
#csv_files = glob.glob('*.csv')

#for filename in csv_files:
    #data = pd.read_csv(filename,encoding='UTF 16-LE', sep='\t')#,dtype={'Stock_Cajas':np.float64, 'Sku_ABInBev': np.float64} ,decimal='.')
  #  SL_KOE.append(data)
#SL_KOE = pd.concat(SL_KOE,ignore_index=True)
#SL_KOE =SL_KOE[["Año_mes_dia_entrega","Codigo_Cliente", "Nom_CD", "SKU_ABInBev", "Cant_cjs_Venta","segmento_marketing"]]

#FL_SOL = FL_KOE[(FL_KOE["ATRIBUTO"]=="MOTIVO_DIF_COSTO")]
#SL_KOE['Motivo']="Stock_out"
#SL_KOE.rename(columns = {'Año_mes_dia_entrega':'Fecha_t','Cant_cjs_Venta': 'Cj','Codigo_Cliente':'COD_CLI','SKU_ABInBev':'SKU','segmento_marketing':'Segmento'} ,inplace=True)

#SL_KOE.Cj = SL_KOE.Cj.astype(float)
#SL_KOE.Cj = SL_KOE.Cj.round(1)
#SL_KOE = SL_KOE[SL_KOE["Cj"]>0]

#SL_KOE.SKU = SL_KOE.SKU.astype(str)
#Titulo = SL_KOE.SKU
#Titulo = Titulo.str.split(".", n=1, expand=True)
#SL_KOE["SKU"]=Titulo[0].str.replace(r".","")
 
#SL_KOE =SL_KOE[["Fecha_t","COD_CLI","Nom_CD","SKU","Cj","Motivo","Segmento"]]
#SL_KOE.to_csv("D:/Reporting/ETL/08 SL Botters/SL_Bottlers.txt")


#################################################### RECHAZO BOTTLER (T2) ########################################################
############# Rechazo KOE ###########

#os.chdir('D:/Reporting/Maestras/')
#Maestro_koe = pd.read_excel('COD KOE SKU.xlsx', sheet_name='Hoja1')

#os.chdir('D:/Reporting/Bases Manuales/12 RECHAZO/KOE')
#Rech_KOE = []
#csv_files = glob.glob('*.csv')

#for filename in csv_files:
 #   data = pd.read_csv(filename,encoding='UTF-16 LE', sep='\t',dtype={'SKU':np.float64,'Cant_cjs_Venta':np.float64,'Cant_Cjs_Rechazo':np.float64,'HL_Rechazo':np.float64} ,decimal='.')
  #  #data['filename']=filename
   # Rech_KOE.append(data)
#Rech_KOE = pd.concat(Rech_KOE,ignore_index=True)
#Rech_KOE

#Rech_KOE.rename(columns = {'Año_mes_dia_entrega':'Fecha_t','Nom_CD':'COD_STOCK', 'SKU': 'Sku_KOE', 'Codigo_Cliente':'COD_CLI','Cant_cjs_Venta':'CJ_VTA','Cant_Cjs_Rechazo':'CJ_RECH','HL_Rechazo':'HL_RECH','COD_Motivo_Movimiento':'MOTIVO'},inplace=True)
#Rech_KOE=Rech_KOE[["Fecha_t","COD_STOCK","Sku_KOE","COD_CLI","MOTIVO","CJ_VTA","CJ_RECH","HL_RECH"]]

#Rech_KOE.CJ_VTA = Rech_KOE.CJ_VTA.round(1)
#Rech_KOE.CJ_RECH = Rech_KOE.CJ_RECH.round(1)
#Rech_KOE.HL_RECH = Rech_KOE.HL_RECH.round(3)

#Rech_KOE = pd.merge(Rech_KOE,Maestro_koe,on=["Sku_KOE"],how="outer")

#Rech_KOE = Rech_KOE.dropna(subset=['Fecha_t'])

#GENERAR MAESTRA DE CLIENTES DESDE SELL OUT
#DISPONIBLE
#MAESTRO DE CONVERCIONES KOE
#Rech_KOE.to_csv("D:/Reporting/ETL/06 Rechazo/Rechazo.txt")
