==> 07 Deployment.ipynb <==
### Tratamiento de datos
import glob
import os
import pandas as pd
import numpy as np
import calendar
from contextlib import contextmanager
from datetime import datetime
import math

########################################## Politica #########################################
os.chdir('D:/Reporting/Bases Manuales/03 POLITICA/Diaria')
Politica = []
csv_files = glob.glob('*.xlsx')

for filename in csv_files:
    data = pd.read_excel(filename,sheet_name='Hoja1', dtype={'FCST HL ajustado':np.float64,'Seg HL':np.float64, 'Repo HL':np.float64, 'Max HL':np.float64, 'OOR Out HL':np.float64, 'OOR Over HL':np.float64} ,decimal=',')
    data['filename']=filename
    Politica.append(data)
Politica = pd.concat(Politica,ignore_index=True)

Titulo = Politica[["filename"]]
Titulo = Titulo["filename"].str.split(".", n=1, expand=True)
Politica["Fecha_t"]=Titulo[0].str.replace(r".","")

Politica.rename(columns ={'FCST HL':'FCST_HL', 'Seg HL': 'Seg_HL', 'Repo HL': 'Repo_HL', 'Max HL': 'Max_HL', 'OOR Out HL': 'OOR_Out_HL', 'OOR Over HL': 'OOR_Over_HL'},inplace=True)
Politica.Seg_HL = np.round(Politica.Seg_HL,1)#Seg_HL
Politica.Repo_HL = np.round(Politica.Repo_HL,1)#Repo_HL
Politica.Max_HL = np.round(Politica.Max_HL,1)#Max_HL
Politica.OOR_Out_HL = np.round(Politica.OOR_Out_HL,1)#OOR_OUT_HL
Politica.OOR_Over_HL = np.round(Politica.OOR_Over_HL,1)#OOR_OVER_HL

Politica = Politica[['Fecha_t', 'Region','Cod. Producto','FCST_HL','Seg_HL', 'Repo_HL', 'Max_HL', 'OOR_Out_HL', 'OOR_Over_HL','Transito Hl','Cargan hoy Hl','D+1  Hl','DV centro','Stock centro hoy HL', 'Sugerido', 'Sugerido revisado', 'Disponibilidad', 'Stock nacional HL', 'Stock Total Hl', 'Stock hoyHL']]

Politica.rename(columns ={'Transito Hl':'Transito','Cargan hoy Hl':'Cargan_hoy','D+1  Hl':'D1','Stock centro hoy HL':'Stock_centro_hoy_HL','DV centro':'DV_centro'},inplace=True)
Politica.rename(columns ={'Stock nacional HL': 'Dispo_ABI_HL', 'Disponibilidad': 'Dispo_HL', 'Stock hoyHL': 'Stock_Tron_HL', 'Stock Total Hl': 'Stock_Py_HL', 'Sugerido': 'SugT_HL', 'Sugerido revisado': 'SugR_HL'},inplace=True)

Politica.Cargan_hoy = np.round(Politica.Cargan_hoy,1)#D0_PLL
Politica.Transito = np.round(Politica.Transito,1)#Transito_PLL
Politica.D1 = np.round(Politica.D1,1)#D1_PLL
Politica.SugT_HL = np.round(Politica.SugT_HL,1)#SugT_HL
Politica.SugR_HL = np.round(Politica.SugR_HL,1)#SugR_HL
Politica.Dispo_HL = np.round(Politica.Dispo_HL,1)#Dispo_HL
Politica.Stock_Py_HL = np.round(Politica.Stock_Py_HL,1)#Stock_Py_HL
Politica.Stock_Tron_HL = np.round(Politica.Stock_Tron_HL,1)#Stock_Tron_HL
Politica.FCST_HL = np.round(Politica.FCST_HL,1)#FCST_HL
Politica.Dispo_ABI_HL = np.round(Politica.Dispo_ABI_HL,1)#Dispo_ABI_HL
Politica.Stock_centro_hoy_HL = np.round(Politica.Stock_centro_hoy_HL,1)# Dispo_Centro_HL 

Politica = Politica.fillna(0)

################################# Nuevo formato ###########################################################
os.chdir('D:/Reporting/Bases Manuales/03 POLITICA/Nuevo formato')
Politica2 = []
csv_files = glob.glob('*.xlsx')

for filename in csv_files:
    data = pd.read_excel(filename,sheet_name='Hoja1', dtype={'FCST_HL':np.float64,'Seg_HL':np.float64, 'Repo_HL':np.float64, 'Max_HL':np.float64, 'OOR_OUT_HL':np.float64, 'OOR_OVER_HL':np.float64} ,decimal=',')
    data['filename']=filename
    Politica2.append(data)
Politica2 = pd.concat(Politica2,ignore_index=True)

Titulo = Politica2[["filename"]]
Titulo = Titulo["filename"].str.split(".", n=1, expand=True)
Politica2["Fecha_t"]=Titulo[0].str.replace(r".","")

Politica2.Seg_HL = np.round(Politica2.Seg_HL,1)#Seg_HL
Politica2.Repo_HL = np.round(Politica2.Repo_HL,1)#Repo_HL
Politica2.Max_HL = np.round(Politica2.Max_HL,1)#Max_HL
Politica2.OOR_OUT_HL = np.round(Politica2.OOR_OUT_HL,1)#OOR_OUT_HL
Politica2.OOR_OVER_HL = np.round(Politica2.OOR_OVER_HL,1)#OOR_OVER_HL

Politica2.D0_PLL = np.round(Politica2.D0_PLL,1)#D0_PLL
Politica2.Transito_PLL = np.round(Politica2.Transito_PLL,1)#Transito_PLL
Politica2.D1_PLL = np.round(Politica2.D1_PLL,1)#D1_PLL
Politica2.SugT_HL = np.round(Politica2.SugT_HL,1)#SugT_HL
Politica2.SugR_HL = np.round(Politica2.SugR_HL,1)#SugR_HL
Politica2.Dispo_HL = np.round(Politica2.Dispo_HL,1)#Dispo_HL
Politica2.Stock_Py_HL = np.round(Politica2.Stock_Py_HL,1)#Stock_Py_HL
Politica2.Stock_Tron_HL = np.round(Politica2.Stock_Tron_HL,1)#Stock_Tron_HL
Politica2.FCST_HL = np.round(Politica2.FCST_HL,1)#FCST_HL
Politica2.Dispo_ABI_HL = np.round(Politica2.Dispo_ABI_HL,1)#Dispo_ABI_HL
Politica2.Dispo_Centro_HL = np.round(Politica2.Dispo_Centro_HL,1)# Dispo_Centro_HL 

Politica2 = Politica2[['Fecha_t', 'Region','Cod. Producto','FCST_HL','Seg_HL', 'Repo_HL', 'Max_HL', 'OOR_OUT_HL', 'OOR_OVER_HL','Transito_PLL','D0_PLL','D1_PLL','Dispo_Centro_HL', 'SugR_HL', 'SugT_HL', 'Dispo_HL', 'Dispo_ABI_HL', 'Stock_Tron_HL', 'Stock_Py_HL']]
Politica2.rename(columns ={'OOR_OUT_HL':'OOR_Out_HL','OOR_OVER_HL':'OOR_Over_HL', 'Transito_PLL': 'Transito', 'D0_PLL':'Cargan_hoy', 'D1_PLL': 'D1', 'Dispo_ABI_HL':'Stock_centro_hoy_HL'},inplace=True)
Politica2 = Politica2.fillna(0)
Politica = Politica.append(Politica2)

#### EXPORTABLE
Politica.to_csv("D:/Reporting/ETL/03 Politica/Politica_Sabana.txt")

### Tratamiento de datos
import glob
import os
import pandas as pd
import numpy as np
import calendar
from contextlib import contextmanager
from datetime import datetime
import math


##### Nuevo formato
os.chdir('D:/Reporting/Bases Manuales/03 POLITICA/Nuevo formato')
Politica2 = []
csv_files = glob.glob('*.xlsx')

for filename in csv_files:
    data = pd.read_excel(filename,sheet_name='Hoja1', dtype={'FCST_HL':np.float64,'Seg_HL':np.float64, 'Repo_HL':np.float64, 'Max_HL':np.float64, 'OOR_OUT_HL':np.float64, 'OOR_OVER_HL':np.float64} ,decimal=',')
    data['filename']=filename
    Politica2.append(data)
Politica2 = pd.concat(Politica2,ignore_index=True)

Titulo = Politica2[["filename"]]
Titulo = Titulo["filename"].str.split(".", n=1, expand=True)
Politica2["Fecha_t"]=Titulo[0].str.replace(r".","")

Politica2.Seg_HL = np.round(Politica2.Seg_HL,1)#Seg_HL
Politica2.Repo_HL = np.round(Politica2.Repo_HL,1)#Repo_HL
Politica2.Max_HL = np.round(Politica2.Max_HL,1)#Max_HL
Politica2.OOR_OUT_HL = np.round(Politica2.OOR_OUT_HL,1)#OOR_OUT_HL
Politica2.OOR_OVER_HL = np.round(Politica2.OOR_OVER_HL,1)#OOR_OVER_HL

Politica2.D0_PLL = np.round(Politica2.D0_PLL,1)#D0_PLL
Politica2.Transito_PLL = np.round(Politica2.Transito_PLL,1)#Transito_PLL
Politica2.D1_PLL = np.round(Politica2.D1_PLL,1)#D1_PLL
Politica2.SugT_HL = np.round(Politica2.SugT_HL,1)#SugT_HL
Politica2.SugR_HL = np.round(Politica2.SugR_HL,1)#SugR_HL
Politica2.Dispo_HL = np.round(Politica2.Dispo_HL,1)#Dispo_HL
Politica2.Stock_Py_HL = np.round(Politica2.Stock_Py_HL,1)#Stock_Py_HL
Politica2.Stock_Tron_HL = np.round(Politica2.Stock_Tron_HL,1)#Stock_Tron_HL
Politica2.FCST_HL = np.round(Politica2.FCST_HL,1)#FCST_HL
Politica2.Dispo_ABI_HL = np.round(Politica2.Dispo_ABI_HL,1)#Dispo_ABI_HL
Politica2.Dispo_Centro_HL = np.round(Politica2.Dispo_Centro_HL,1)# Dispo_Centro_HL 

Politica2 = Politica2[['Fecha_t', 'Region','Cod. Producto','FCST_HL','Seg_HL', 'Repo_HL', 'Max_HL', 'OOR_OUT_HL', 'OOR_OVER_HL','Transito_PLL','D0_PLL','D1_PLL','Dispo_Centro_HL', 'SugR_HL', 'SugT_HL', 'Dispo_HL', 'Dispo_ABI_HL', 'Stock_Tron_HL', 'Stock_Py_HL']]
Politica2.rename(columns ={'OOR_OUT_HL':'OOR_Out_HL','OOR_OVER_HL':'OOR_Over_HL', 'Transito_PLL': 'Transito', 'D0_PLL':'Cargan_hoy', 'D1_PLL': 'D1', 'Dispo_ABI_HL':'Stock_centro_hoy_HL'},inplace=True)


Politica2.info()

os.chdir('D:/Reporting/Bases Manuales/10 ADH/04 Progama nuevo')

Programa = []
csv_files = glob.glob('*.xlsx')

for filename in csv_files:
    data = pd.read_excel(filename,sheet_name='detalle de cargas',decimal=',')#, dtype={'FCST HL ajustado':np.float64,'Seg HL':np.float64, 'Repo HL':np.float64, 'Max HL':np.float64, 'OOR Out HL':np.float64, 'OOR Over HL':np.float64} ,decimal=',')
    #data['filename']=filename
    Programa.append(data)
Programa = pd.concat(Programa,ignore_index=True)
Programa = Programa[["Fecha", "CD DESTINO", "DT","SKU TRUCK", "Pallets", "Hls"]]
Programa = Programa[Programa["Hls"]!=' ']

Programa = Programa.dropna(subset=['Hls'])
Programa = Programa.fillna(0)
Programa.Hls = Programa.Hls.astype(float)
Programa.Hls = np.round(Programa.Hls,1)

Programa.to_csv("D:/Reporting/ETL/03 Politica/Programa.txt")
Programa.info()

Politica = Politica[['Fecha_t', 'Region','Cod. Producto','FCST_HL','Seg_HL', 'Repo_HL', 'Max_HL', 'OOR_Out_HL', 'OOR_Over_HL','Transito Hl','Cargan hoy Hl','D+1  Hl','DV centro','Stock centro hoy HL', 'Sugerido', 'Sugerido revisado', 'Disponibilidad', 'Stock nacional HL', 'Stock Total Hl', 'Stock hoyHL']]
Politica.rename(columns ={'Transito Hl':'Transito','Cargan hoy Hl':'Cargan_hoy','D+1  Hl':'D1','Stock centro hoy HL':'Stock_centro_hoy_HL','DV centro':'DV_centro'},inplace=True)
Politica.rename(columns ={'Stock nacional HL': 'Dispo_ABI_HL', 'Disponibilidad': 'Dispo_HL', 'Stock hoyHL': 'Stock_Tron_HL', 'Sugerido': 'Sugerido revisado', 'Stock Total Hl': 'Stock_Py_HL'},inplace=True)

Politica.Cargan_hoy = np.round(Politica.Cargan_hoy,1)
Politica.Transito = np.round(Politica.Transito,1)
Politica.Stock_centro_hoy_HL = np.round(Politica.Stock_centro_hoy_HL,1)

#Politica.DV_centro = np.round(Politica.DV_centro,1)
Politica.D1 = np.round(Politica.D1,1)

### Tratamiento de datos
import glob
import os
import pandas as pd
import numpy as np
import calendar
from contextlib import contextmanager
from datetime import datetime
import math

###############################################################################################################################################
#### Produccion Plan
os.chdir('D:/BI Plannig/01 Imput/02 Plan Produccion')
P_P = []
csv_files = glob.glob('*.xlsx')

for filename in csv_files:
    data = pd.read_excel(filename,sheet_name='Input marito')#,dtype={'Sugerido D+2':np.float64,'Sugerido revisado D + 2':np.float64})
    P_P.append(data)
P_P = pd.concat(P_P,ignore_index=True)

P_P = P_P.drop(['Descripción'], axis=1)
P_P = P_P.dropna(subset=['Material '])
P_P = P_P.set_index(('Material '))

## Transformar en una entrada
sr = P_P.stack()
st = pd.DataFrame(sr).reset_index()
st['level_1'] = st.level_1.astype(str)
t = st.level_1.str.split(" ", n=1, expand=True)
st["Fecha"] = t[0]
st = st[st["Fecha"]!="sem"]
st.rename(columns = {0:'Valor','Material ':'Material'} ,inplace=True)
PP_ok = st[['Material','Fecha','Valor']]

### SKU
PP_ok.Material  = PP_ok.Material.astype(str)
Titulo = PP_ok[["Material"]]
Titulo = Titulo["Material"].str.split(".", n=1 , expand=True)
PP_ok['Material'] =  Titulo[0].astype(str)
PP_ok.Valor = PP_ok.Valor.astype(str)
PP_ok["Valor"] = PP_ok.Valor.str.replace(r"º","")

### Fecha
PP_ok.Fecha  = PP_ok.Fecha.astype(str)
Titulo = PP_ok[["Fecha"]]
Titulo = Titulo["Fecha"].str.split("-", n=2 , expand=True)
PP_ok['Fecha'] =  Titulo[0].astype(str) + Titulo[1].astype(str) + Titulo[2].astype(str)

PP_ok.Valor = PP_ok.Valor.astype(float)
PP_ok.Valor = PP_ok.Valor.round(3)
PP_ok.Valor = PP_ok.Valor*1000
PP_ok.Valor = PP_ok.Valor.round(0)
PP_ok = PP_ok[PP_ok["Valor"]>0]

PP_ok.to_csv("D:/BI Plannig/02 ETL/02 DATA/03 Produccion Plan/Produccion_Plan.txt")

#####################################################################################################################################################


################################################################################################################################################

##### Maquila Plan
os.chdir('D:/BI Plannig/01 Imput/04 Plan de maquila')
M_P = []
csv_files = glob.glob('*.xlsx')

for filename in csv_files:
    data = pd.read_excel(filename,sheet_name='Input marito')#,dtype={'Sugerido D+2':np.float64,'Sugerido revisado D + 2':np.float64})
    M_P.append(data)
M_P = pd.concat(M_P,ignore_index=True)

M_P = M_P.drop(['SAP'], axis=1)
M_P = M_P.drop(['Producto'], axis=1)
M_P = M_P.dropna(subset=['Truck'])
M_P = M_P.set_index(('Truck'))

## Transformar en una entrada
sr = M_P.stack()
st = pd.DataFrame(sr).reset_index()
st.level_1 = st.level_1.astype(str)
t = st.level_1.str.split(" ", n=1, expand=True)
st["CP"] = t[0]
st = st[st["CP"]!="sem"]
st = st[st[0]>0]
st.rename(columns = {0:'Valor','Truck':'Material','CP':'Fecha'} ,inplace=True)
MP_ok = st[['Material','Fecha','Valor']]

MP_ok.Valor = MP_ok.Valor.astype(float)
MP_ok.Valor = MP_ok.Valor.round(3)
MP_ok.Valor = MP_ok.Valor*1000
MP_ok.Valor = MP_ok.Valor.round(0)

MP_ok.Material = MP_ok.Material.astype(str)

### Fecha
MP_ok.Fecha  = MP_ok.Fecha.astype(str)
Titulo = MP_ok[["Fecha"]]
Titulo = Titulo["Fecha"].str.split("-", n=2, expand=True)
MP_ok['Fecha'] =  Titulo[0].astype(str) + Titulo[1].astype(str) + Titulo[2].astype(str)

MP_ok.to_csv("D:/BI Plannig/02 ETL/02 DATA/04 Maquila Plan/Maquila_Plan.txt")
